backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.2
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
websockets==12.0
google-cloud-speech==2.21.0
vosk==0.3.45
pydantic==2.5.0
httpx==0.25.2
python-socketio==5.10.0
eventlet==0.33.3
